// src/app/core/models/role.ts
export enum Role {
  All = 'All',
  MasterAdmin = 'Master Admin',
  SuperAdmin = 'Super Admin',
  JSTISchool = 'JSTI School',
  BNASchool = 'BNA School',
  Student = 'Student',
  Instructor = 'Instructor',
  InterSeeviceCourse = 'Inter Seevice Course',
  InterSeeviceDesk = 'Inter Seevice Desk',
  BNA = 'BNA',
  ForeignTraining='Foreign Training',
  DDNT='DDNT',
  DataEntry = 'Data Entry',
  BnaDataEntry= 'BNA Data Entry',
  CO='CO',
  TrainingOffice='Training Office',
  SchoolOIC='School OIC',
  TC='TC',
  AreaCommander = 'Area Commander',
  Director ='Director',
  TCO ='TCO',
  OICNBCDSchool='OIC NBCD School',
  OICNETSchool='OIC NET School'
}